[COS-1611] small poc without proper data
