--- packages/server/customer-os-neo4j-repository/repository/godogs.go
@@ -1,6 +1,0 @@
-package repository
-
-// Godogs available to eat
-var Godogs int
-
-func main() { /* usual main func */ }
